% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAge.R
\name{compute_age}
\alias{compute_age}
\title{Compute age}
\usage{
compute_age(start_date, end_date = Sys.Date())
}
\arguments{
\item{start_date}{The origin date}

\item{end_date}{The end date.}
}
\description{
Compute the age in years based on a start date and a end date
}
\examples{
\dontrun{

library(toolboxDekuple)

# With default end_date
compute_age(start_date = "2020-01-01")

# With specific end_date
compute_age(start_date = "2020-01-01", end_date = "2024-01-01")
}

}
