% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMarginalFrequency.R
\name{compute_marginal_frequency}
\alias{compute_marginal_frequency}
\title{Compute marginal frequencies}
\usage{
compute_marginal_frequency(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
Compute the sum by row and column of the numeric columns
}
